+++ scripts/release-notes-ai-generator/prompts/generation.md
