+++ scripts/release-notes-ai-generator/requirements.txt
